backend almost done

+++ controllers/orders.js
@@ -2,19 +2,18 @@ const { validationResult } = require('express-validator');
 const Orders = require('../models/Orders');
 
 
-// CREATE ORDER
+// PLACE ORDER
 
 async function makeOrder(req, res) {
     try {
-        const errors = validationResult(req);
-        // Checking if validations are fulfilled
-        if (!errors.isEmpty()) {
-            return res.status(400).json({ errors: errors.array() });
-        }
         const order = new Orders({
             user: req.user.id,
-            product: req.params.id,
-            status: req.body.status
+            products: req.body.products,
+            status: req.body.status,
+            address: req.body.address,
+            contact: req.body.contact,
+            postal_code: req.body.postal_code,
+            city: req.body.city
         })
         const savedOrder = await order.save();
         res.json(savedOrder);
@@ -26,7 +25,7 @@ async function makeOrder(req, res) {
 
 
 
-// RETRIEVE PRODUCTS
+// RETRIEVE ALL ORDERS
 
 async function getOrders(req, res) {
     try {
@@ -38,9 +37,21 @@ async function getOrders(req, res) {
     }
 }
 
+// RETRIEVE ORDERS OF SPECIFIC USER
+
+async function getOrder(req, res) {
+    try {
+        const orders = await Orders.find({ user: req.user.id })
+        res.json(orders);
+    } catch (error) {
+        console.log(error);
+        res.status(500).send("Internal server error")
+    }
+}
 
 
-// UPDATE PRODUCT
+
+// UPDATE ORDER ADMIN SIDE
 
 async function updateOrder(req, res) {
     try {
@@ -58,14 +69,33 @@ async function updateOrder(req, res) {
 }
 
 
-// DELETE PRODUCT
+// DELETE ORDER ADMIN SIDE
 
 async function deleteOrder(req, res) {
     try {
         let order = await Orders.findById(req.params.id);
         if (!order) { return res.status(404).send("Not found!") }
+        order = await Orders.findByIdAndDelete(req.params.id);
+        res.json({ "Success": "Order has been deleted.", order: order });
+    } catch (error) {
+        console.log(error);
+        res.status(500).send("Internal server error")
+    }
+}
+
+
+// DELETE ORDER USER SIDE
+
+async function deleteOrderUser(req, res) {
+    try {
+        let order = await Orders.findById(req.params.id);
+        if (!order) { return res.status(404).send("Not found!") }
+        if (order.user.toString() == req.user.id) {
             order = await Orders.findByIdAndDelete(req.params.id);
             res.json({ "Success": "Order has been deleted.", order: order });
+        } else if (item.user.toString() != req.user.id) {
+            return res.status(401).send("Not allowed!")
+        }
     } catch (error) {
         console.log(error);
         res.status(500).send("Internal server error")
@@ -75,8 +105,10 @@ async function deleteOrder(req, res) {
 
 module.exports = {
     getOrders,
+    getOrder,
     makeOrder,
     updateOrder,
-    deleteOrder
+    deleteOrder,
+    deleteOrderUser
 }
